01_OAS/requirements.txt
beautifulsoup4
Crawl4AI
openai
playwright
pyairtable
pypdf
python-dotenv
requests
supabase
tiktoken
